+++ packages/dart_agents/test/mini_swe_agent_test.dart
@@ -1,5 +1,5 @@
 import 'package:ai_primitives/ai_primitives.dart';
-import 'package:mini_swe_agent/mini_swe_agent.dart';
+import 'package:dart_agents/dart_agents.dart';
 import 'package:test/test.dart';
 
 /// A fake [AI] that replays a scripted sequence of responses, one per